Omit tag <values> in output, for a better compatibility w/ XMCDA v2 prgs
Most XMCDA v2 programs do not use the container tag <values>.  While
this should not be an issue, some of them do not expect <values> as
inputs and fail if it is present.  Hence, it is safer to omit it in
the outputs.

NB: XMCDA v3 does not suffer from such a problem since the container
    tag <values> is mandatory.